LayoutTests/imported/w3c/web-platform-tests/css/css-env/env-in-custom-properties.tentative-expected.txt


PASS env() is substituted into a custom property
PASS Substitution of unrecognized env() causes guaranteed-invalid
